requirements.txt
# FastAPI & Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Qdrant Vector Database
qdrant-client>=1.7.0

# OpenAI API
openai>=1.3.0

# Redmine API
python-redmine>=2.3.0

# Streamlit UI
streamlit>=1.28.0

# Data & Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Progress bar for scripts
tqdm>=4.66.1

# HTTP client
httpx>=0.25.2
requests>=2.31.0

# Date/Time utilities
python-dateutil>=2.8.2
